--- components/form/demo/FormModal/index.ts
@@ -1,5 +1,0 @@
-import Action from './Action';
-import FormModal from './FormModal';
-
-export { Action };
-export default FormModal;
