+++ components/form/demo/form-in-modal.tsx
@@ -1,7 +1,120 @@
-import React, { useState } from 'react';
-import { Button, Form, Input, message, Radio } from 'antd';
+import React, { createContext, useContext, useEffect, useMemo, useState } from 'react';
+import type { FormProps, ModalProps } from 'antd';
+import { Button, Form, Input, message, Modal, Radio } from 'antd';
 
-import FormModal, { Action } from './FormModal';
+// ------------------------------------------------------------------FormModal start-----------------------------------------------------------
+
+export type onCancel = (e?: React.MouseEvent<HTMLElement>) => void;
+
+export interface ActionContextProps {
+  open?: boolean;
+  onCancel: onCancel;
+  onDestroy: () => void;
+}
+export const Context = createContext<ActionContextProps>({
+  onCancel: () => undefined,
+  onDestroy: () => undefined,
+});
+
+export interface ActionProps extends Pick<ActionContextProps, 'open' | 'onCancel'> {
+  children?: React.ReactNode;
+}
+
+const Action = (props: ActionProps) => {
+  const { onCancel, children, open: propsOpen } = props;
+
+  const [load, setLoad] = useState<Record<string, any>>();
+  const [open, setOpen] = useState<boolean>();
+
+  useEffect(() => {
+    if (propsOpen) {
+      setLoad({});
+    } else {
+      setOpen(false);
+    }
+  }, [propsOpen]);
+
+  useEffect(() => {
+    setOpen(!!load);
+  }, [load]);
+
+  const context = useMemo<ActionContextProps>(
+    () => ({
+      open,
+      onCancel: (e) => {
+        onCancel?.(e);
+        setOpen(false);
+      },
+      onDestroy: () => setLoad(undefined),
+    }),
+    [open],
+  );
+
+  return <Context.Provider value={context}>{(load || propsOpen) && children}</Context.Provider>;
+};
+
+export const ContextReset = ({ children }: { children: React.ReactNode }) => {
+  const context = useMemo<ActionContextProps>(
+    () => ({ onCancel: () => undefined, onDestroy: () => undefined }),
+    [],
+  );
+  return <Context.Provider value={context}>{children}</Context.Provider>;
+};
+
+export interface FormModalProps<Values = any> extends ModalProps {
+  children?: React.ReactNode;
+  onCancel?: onCancel;
+  initialValues?: FormProps<Values>['initialValues'];
+  onFinish?: FormProps<Values>['onFinish'];
+  formProps?: Omit<FormProps<Values>, 'initialValues' | 'onFinish'>;
+}
+
+function FormModal<T = any>(props: FormModalProps<T>) {
+  const { onDestroy, ...contextRest } = useContext(Context);
+  const {
+    children,
+    onCancel,
+    initialValues,
+    onFinish,
+    formProps = {},
+    ...rest
+  } = { ...contextRest, ...props };
+  const [loading, setLoading] = useState(false);
+
+  const handleOnFinish = async (values: T) => {
+    if (loading) return;
+    setLoading(true);
+    try {
+      await onFinish?.(values);
+      onCancel?.();
+      setLoading(false);
+    } catch (error) {
+      setLoading(false);
+    }
+  };
+
+  const modalRender: ModalProps['modalRender'] = (node) => (
+    <Form initialValues={initialValues} onFinish={handleOnFinish} {...formProps}>
+      {node}
+    </Form>
+  );
+
+  return (
+    <Modal
+      modalRender={modalRender}
+      afterClose={() => onDestroy?.()}
+      onCancel={onCancel}
+      destroyOnClose
+      cancelButtonProps={{ onClick: onCancel }}
+      okButtonProps={{ htmlType: 'submit', loading }}
+      {...rest}
+    >
+      <ContextReset>{children}</ContextReset>
+    </Modal>
+  );
+}
+
+// ------------------------------------------------------------------FormModal end-----------------------------------------------------------
 
 const CollectionCreateForm = () => {
   const [form] = Form.useForm();
